+++ crates/riscv-core/src/core/execute/rv32i.rs
@@ -95,14 +95,15 @@ impl Cpu {
             Lhu => (false, 2),
             _   => return None,
         };
-        let mode = self.mode;
-        let bus = &mut self.bus;
 
-        Some(
-            if is_signed {
-                Lsu::load_signed(bus, &self.csrs, src, offset, byte_num, mode)
+        let mut lsu = Lsu::new(
+            &mut self.mmu, &mut self.bus, &self.csrs, self.mode
+        );
+ 
+        Some(if is_signed {
+                lsu.load_signed(src, offset, byte_num)
             } else {
-                Lsu::load(bus, &self.csrs, src, offset, byte_num, mode)
+                lsu.load(src, offset, byte_num)
             }
         )
     }
@@ -114,12 +115,11 @@ impl Cpu {
             Sw => 4,
             _  => return None,
         };
-        let mode = self.mode;
-        let bus = &mut self.bus;
+        let mut lsu = Lsu::new(
+            &mut self.mmu, &mut self.bus, &self.csrs, self.mode
+        );
 
-        Some(
-            Lsu::store(bus,  &self.csrs, des, src, offset, byte_num, mode)
-        )
+        Some(lsu.store(des, src, offset, byte_num))
     }
 
     fn branch(op: Rv32iOp, data1: u32, data2: u32) -> Option<bool> {