+++ crates/riscv-core/src/core/mmu/tlb/result.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum TlbResult {
+    Hit(bool, u32),
+    Miss,
+    PageFault,
+    UpdateAD,
+}